fix(k8s): Skip HPA creation when min and max replicas are equal
The Horizontal Pod Autoscaler should only be created when there's actually
a scaling range (when maxReplicas > minReplicas). Previously, we were
creating HPAs even when these values were equal, which is unnecessary
since no actual scaling can occur in that case.

This also provides a more elegant solution to remove the HPA. I stupidly
thought that we can remove the HPA if we remove the `maxReplicas` and
`minReplicas` arguments (theplant/lacoste-provisioning#2185), but it's
not true.